Move path breadcrumb to separate control
